+++ docs/source/change_log.rst
@@ -15,6 +15,9 @@ Version 4.0.0
 * **Breaking:** Changed ``STRING.to_ary`` to return an array of characters instead of splitting the string
     * Use the new builtin ``$split`` function to split a string on whitespace into an array of words
 * **Breaking:** Changed :py:class:`~rule_engine.engine.Context` to use keyword-only arguments
+* **Breaking:** Dropped support for Python versions 3.4 and 3.5
+* **Breaking:** Invalid floating point literals now raise :py:exc:`~.errors.FloatSyntaxError` instead of
+  :py:exc:`~.errors.RuleSyntaxError`
 
 Version 3.x.x
 -------------
@@ -24,7 +27,7 @@ Version 3.6.0
 
 Released :release:`3.6.0` on June 16th, 2023
 
-* Removed testing for Python 3.4 and 3.5 on GitHub Actions
+* Removed testing for Python versions 3.4 and 3.5 on GitHub Actions
 * Add regex error details to the debug REPL
 * Add support for Python-style comments